requirements.txt
cramjam==2.5.0
fastparquet==0.8.1
fsspec==2022.5.0
greenlet==1.1.1
joblib==1.0.1
numpy==1.21.1
pandas==1.3.1
patsy==0.5.1
psycopg2==2.9.1
pyarrow==8.0.0
pymssql==2.2.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.22
statsmodels==0.12.2
tqdm==4.64.0
